requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
django-cors-headers==4.7.0
django-email-users==2.1.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework-jsonapi==7.1.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
inflection==0.5.1
kombu==5.5.4
oauthlib==3.3.1
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
